% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneSample_2ScML.R
\name{OneSample_2ScML}
\alias{OneSample_2ScML}
\title{2ScML with One Sample}
\usage{
OneSample_2ScML(Y, D, Z, K1_vec, K2_vec, tau1 = 1e-05, tau2 = 1e-05)
}
\arguments{
\item{Y}{Response vector of length n.}

\item{D}{Exposure vector of length n.}

\item{Z}{IV matrix, n rows and p columns, p columns correspond to p IVs.}

\item{K1_vec}{Vector of candidate K1's for stage 1.}

\item{K2_vec}{Vector of candidate K2's for stage 2.}

\item{tau1}{Parameter tau in TLC for stage 1.}

\item{tau2}{Parameter tau in TLC for stage 2.}
}
\value{
A list contains following elements.
K1: BIC selected K1 in stage 1;
K2: BIC selected K2 in stage 2;
stage1_ind: indices of IVs selected as relevant in stage 1;
stage2_ind: indices of IVs selected as invalid in stage 2;
beta_est: estimate of causal effect of exposure to outcome;
beta_se: standard error of beta_est.
}
\description{
Main function perform 2ScML with one sample,
using BIC for model selection in both stages.
}
