% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneSampleEstSE.R
\name{OneSampleEstSE}
\alias{OneSampleEstSE}
\title{One Sample Estimate and Standard Error of Causal Effect}
\usage{
OneSampleEstSE(Y, D, Z, stage1_ind, stage2_ind)
}
\arguments{
\item{Y}{Response vector of length n.}

\item{D}{Exposure vector of length n.}

\item{Z}{IV matrix, n rows and p columns, p columns correspond to p IVs.}

\item{stage1_ind}{Indices of IVs selected as relevant in stage 1.}

\item{stage2_ind}{Indices of IVs selected as invalid in stage 2.}
}
\value{
A vector of two elements,
the first one is estimated beta,
the second one is its standard error.
}
\description{
For one sample case, given selected relevant IVs in stage 1 and
selected invalid IVs in stage 2,
get estimate of beta, which is the causal effect from exposure to outcome,
and its standard error.
}
