% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetLM.R
\name{SubsetLM}
\alias{SubsetLM}
\title{Fit Linear Model on A Subset of Design Matrix}
\usage{
SubsetLM(Y, X, ind)
}
\arguments{
\item{Y}{Response vector.}

\item{X}{Full design matrix.}

\item{ind}{Set of columns of X to be used in the linear model.}
}
\value{
beta: a vector of estimated coefficients, with non-zero elements for variables being
used and 0's for others.
}
\description{
This function is for fitting a linear model of response Y on a subset of design matrix
X.
}
