% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TScMLStage1.R
\name{TScMLStage1}
\alias{TScMLStage1}
\title{Fit First Stage Model for 2ScML}
\usage{
TScMLStage1(cor.D1Z1, Cap.Sigma.stage1, n1, p, ind.stage1, posdef = 1e-07)
}
\arguments{
\item{cor.D1Z1}{A vector of length p, containing sample correlations
(calculated with the first sample) between exposure D and p instruments.}

\item{Cap.Sigma.stage1}{Sample correlation matrix between p instruments,
could be calculated either with the first sample individual-level instruments data,
or with the reference panel.}

\item{n1}{First stage sample size.}

\item{p}{Number of instruments.}

\item{ind.stage1}{Indices of instruments relevant to exposure D.}

\item{posdef}{A small positive constant to be added on diagonal of covariance matrix
to ensure it is positive definite.}
}
\value{
A vector of length p, containing estimated effects from IVs to exposure D.
}
\description{
This function is used for fitting first stage model of 2ScML.
}
