% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TScMLStage2.R
\name{TScMLStage2}
\alias{TScMLStage2}
\title{Fit Second Stage of 2ScML}
\usage{
TScMLStage2(
  gamma.hat.stage1,
  cor.Y2Z2,
  Estimated.Sigma,
  n1,
  n2,
  p,
  K.vec.stage2 = 0:(p/2),
  posdef = 1e-04,
  Est.Sigma1Square,
  Est.Sigma2Square
)
}
\arguments{
\item{gamma.hat.stage1}{A vector of length p, containing estimated effects from IVs to exposure
obtained in the first stage model.}

\item{cor.Y2Z2}{A vector of length p, containing sample correlations (summary data) between
outcome Y and p IVs calculated with the second sample.}

\item{Estimated.Sigma}{Estimated correlation matrix with reference panel.}

\item{n1}{Sample size in first stage.}

\item{n2}{Sample size in second stage.}

\item{p}{Number of instruments.}

\item{K.vec.stage2}{Set of candidate K's being used in the constraint.}

\item{posdef}{A small positive constant to be added on diagonal of covariance matrix
to ensure it is positive definite.}

\item{Est.Sigma1Square}{A number, estimated variance of random error in the first stage.}

\item{Est.Sigma2Square}{A number, estimated variance of random error in the second stage.}
}
\value{
This function returns a list with three elements,
(1) betaalpha.hat.stage2: a vector of length (p+1), the first element is the estimated beta
(causal effect from exposure to outcome), the rest p elements are estimated alpha's
(direct effects from IVs to outcome);
(2) Asympt.Var.BetaHat: the uncorrected estimated variance for estimated beta.
(3) BIC.vec.stage2: the vector of BIC, having the same length as K.vec.stage2.
}
\description{
This function fitting the second stage of 2ScML is the main function of our proposed method.
It applies Truncated L1 Constraint in the second stage regression to select invalid instruments.
}
