% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TScMLVar.R
\name{TScMLVar}
\alias{TScMLVar}
\title{Corrected Variance for Estimated Causal Effect}
\usage{
TScMLVar(
  Z.ref.original,
  Stage1FittedModel,
  betaalpha.hat.stage2,
  Est.Sigma1Square,
  Est.Sigma2Square,
  n1,
  n2,
  n.ref
)
}
\arguments{
\item{Z.ref.original}{The original reference panel, which is a n.ref (sample size of reference panel)
by p (number of instruments) matrix.}

\item{Stage1FittedModel}{A vector of length p, containing estimated effects from IVs to exposure
obtained in the first stage model.}

\item{betaalpha.hat.stage2}{A vector of length (p+1), the first element is the estimated beta
(causal effect from exposure to outcome), the rest p elements are estimated alpha's
(direct effects from IVs to outcome).}

\item{Est.Sigma1Square}{A number, estimated variance of random error in the first stage.}

\item{Est.Sigma2Square}{A number, estimated variance of random error in the second stage.}

\item{n1}{Sample size in first stage.}

\item{n2}{Sample size in second stage.}

\item{n.ref}{Sample size of reference panel.}
}
\value{
A number, which is the estimated corrected variance of estimated causal effect.
}
\description{
This function is for estimating corrected variance of estimated causal effect (i.e. beta)
obtained from second stage. As described in our draft, the correction accounts for the
effect of using reference panel.
}
