% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSample_2ScML.R
\name{TwoSample_2ScML}
\alias{TwoSample_2ScML}
\title{2ScML with Two Sample}
\usage{
TwoSample_2ScML(
  Y,
  Z,
  stage1_ind,
  gamma_hat,
  Theta0_hat,
  sigma2_hat,
  n2,
  K2_vec,
  tau2 = 1e-05
)
}
\arguments{
\item{Y}{Response vector of length n.}

\item{Z}{IV matrix, n rows and p columns, p columns correspond to p IVs.}

\item{stage1_ind}{Indices of relevant IVs in the first stage.}

\item{gamma_hat}{Estimated gamma in the first stage.}

\item{Theta0_hat}{Asymptotic covariance matrix of gamma_hat as in Assumption 5.}

\item{sigma2_hat}{Estimated sigma2 square as in Assumption 5.}

\item{n2}{Sample size of the second sample used for estimating gamma.}

\item{K2_vec}{Vector of candidate K2's for stage 2.}

\item{tau2}{Parameter tau in TLC for stage 2.}
}
\value{
A list contains following elements.
K2: BIC selected K2 in stage 2;
stage2_ind: indices of IVs selected as invalid in stage 2;
beta_est: estimate of causal effect of exposure to outcome;
beta_se: standard error of beta_est.
}
\description{
Main function perform 2ScML with two sample,
using BIC for model selection in the second stage.
}
