% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleEstSE.R
\name{TwoSampleEstSE}
\alias{TwoSampleEstSE}
\title{Two Sample Estimate and Standard Error of Causal Effect}
\usage{
TwoSampleEstSE(
  Y,
  Z,
  stage1_ind,
  gamma_hat,
  Theta0_hat,
  sigma2_hat,
  n2,
  stage2_ind
)
}
\arguments{
\item{Y}{Response vector of length n.}

\item{Z}{IV matrix, n rows and p columns, p columns correspond to p IVs.}

\item{stage1_ind}{Indices of IVs selected as relevant in stage 1.}

\item{gamma_hat}{Estimated gamma in the first stage.}

\item{Theta0_hat}{Asymptotic covariance matrix of gamma_hat as in Assumption 5.}

\item{sigma2_hat}{Estimated sigma2 square as in Assumption 5.}

\item{n2}{Sample size of the second sample used for estimating gamma.}

\item{stage2_ind}{Indices of IVs selected as invalid in stage 2.}
}
\value{
A vector of two elements,
the first one is estimated beta,
the second one is its standard error.
}
\description{
Two Sample Estimate and Standard Error of Causal Effect
}
